+++ src/CiPeWorldCup/DTO/MatchDTO.cs
@@ -4,5 +4,5 @@ public class MatchDTO
 {
     public int Id { get; set; }
     public int RoundNumber { get; set; }
-    public IEnumerable<long> ParticipantIds { get; set; } = new List<long>();
+    public IEnumerable<int> ParticipantIds { get; set; } = new List<int>();
 }
